testing/web-platform/tests/atomics/window-and-any-worker/wake/wake-nan/wake-nan.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
  <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

// Copyright (C) 2017 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wake wakes zero waiters if the count is NaN
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    let worker = new Worker('wake-nan.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, "timed-out");
    });
    let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    worker.postMessage([ia.buffer]);
    setTimeout(() => assert_equals(0, Atomics.wake(ia, 0, NaN)), 500); // Don't actually wake it
  }, "atomics-dedicated-worker-wake-nan");
  async_test(function(t) {
    let worker = new Worker('wake-nan-worker.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, "timed-out");
    });
    worker.postMessage([]);
  }, "atomics-dedicated-worker-wake-nan-from-worker");
  async_test(function(t) {
    let worker = new SharedWorker('wake-nan-shared-worker.js');
    worker.port.start();
    worker.port.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, "timed-out");
    });
    worker.port.postMessage([]);
  }, "atomics-dedicated-worker-wake-nan-from-shared-worker");
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let test = this;
    contentWindow.onmessage = function(e) {
      let worker = new Worker('wake-nan.js');
      worker.onmessage = test.step_func_done(function(e) {
        assert_equals(e.data, "timed-out");
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia.buffer]);
      setTimeout(() => assert_equals(0, Atomics.wake(ia, 0, NaN)), 500); // Don't actually wake it
    }
    contentWindow.postMessage([], '*');
  }, "atomics-dedicated-worker-wake-nan-from-iframe");
});

</script>
